Fix failing ArtisticEffect test
